backend/pyproject.toml
# FairPact Backend - Python Project Configuration

[project]
name = "fairpact-backend"
version = "0.1.0"
description = "FairPact contract analysis backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    # Document processing
    "python-docx==1.1.0",
    "PyMuPDF==1.23.8",  # Faster and better structure preservation than pdfplumber
    "pytesseract==0.3.10",
    "Pillow==10.2.0",
    # ML/NLP
    "sentence-transformers>=3.0.0,<4.0.0",
    "torch>=2.2.0,<3.0.0",
    "numpy>=1.26.0,<3.0.0",
    # Database
    "sqlalchemy[asyncio]==2.0.25",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pgvector==0.2.4",
    # Storage
    "minio==7.2.3",
    # Async tasks
    "celery==5.3.6",
    "redis==5.0.1",
    # Validation
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "isort==5.13.2",
    "mypy==1.8.0",
    "flake8==7.0.0",
    "pylint==3.0.3",
    "pytest==7.4.3",
    "pytest-asyncio==0.23.2",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "httpx==0.26.0",
    "faker==22.0.0",
    "factory-boy==3.3.0",
    "ipython==8.19.0",
    "ipdb==0.13.13",
    "python-dotenv==1.0.0",
    "watchdog==3.0.0",
    "types-redis==4.6.0.11",
    "types-pillow==10.1.0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "database*", "schemas*", "services*", "migrations*", "models*", "tasks*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["sentence_transformers.*", "pgvector.*", "fitz", "pytesseract.*", "minio.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=. --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
